Improved pagination, added modal to delete laureate/prize

+++ src/app/laureate.component.ts
@@ -36,9 +36,11 @@ export class LaureateComponent {
     }
 
     deleteLaureate(laureateID: string) {
-        this.webService.deleteLaureate(laureateID).subscribe((response: any) => {
-            return this.laureates_list = this.webService.getLaureates(this.page);
-        });
+        if (confirm("Are you sure you want to delete this laureate? This action is irreversible.")) {
+            this.webService.deleteLaureate(laureateID).subscribe((response: any) => {
+                return this.laureates_list = this.webService.getLaureates(this.page);
+            });
+        }
     }
 
     isIncomplete() {
